represent c as int128 in storage, integrate c calculation into pool functions

+++ contracts/pool/PoolStorage.sol
@@ -11,7 +11,8 @@ library PoolStorage {
     address pair;
     address base;
     address underlying;
-    uint c;
+    // fixed point representation
+    int128 c;
   }
 
   function layout () internal pure returns (Layout storage l) {
